Combined NextButton and PrevButton into a single file to share imports and selectors
